refactor(core): improve module structure
