+++ packages/core/src/component/component-collector-legacy.ts
@@ -1,10 +1,56 @@
 import { getClassIdentifier, NodeType, type TSESTreeClass } from "@eslint-react/ast";
+import { getPragmaFromContext } from "@eslint-react/jsx";
 import { type RuleContext, uid } from "@eslint-react/shared";
-import { E, O } from "@eslint-react/tools";
-import type { ESLintUtils } from "@typescript-eslint/utils";
+import { E, M, O } from "@eslint-react/tools";
+import type { ESLintUtils, TSESTree } from "@typescript-eslint/utils";
 
+import type { ExRClassComponent } from "./component";
 import { ExRClassComponentFlag } from "./component-flag";
-import { type ExRClassComponent, isClassComponent, isPureComponent } from "./component-kind";
+
+/**
+ * Check if a node is a React class component
+ * @param node The AST node to check
+ * @param context The rule context
+ */
+export function isClassComponent(node: TSESTree.Node, context: RuleContext): node is TSESTreeClass {
+  if (!("superClass" in node && node.superClass)) {
+    return false;
+  }
+  const pragma = getPragmaFromContext(context);
+  const { superClass } = node;
+
+  return M.match(superClass)
+    .with({ type: NodeType.Identifier, name: M.P.string }, ({ name }) => /^(Pure)?Component$/u.test(name))
+    .with(
+      {
+        type: NodeType.MemberExpression,
+        object: { name: pragma },
+        property: { name: M.P.string },
+      },
+      ({ property }) => /^(Pure)?Component$/u.test(property.name),
+    )
+    .otherwise(() => false);
+}
+
+/**
+ * Check if a node is a React PureComponent
+ * @param node The AST node to check
+ * @param context The rule context
+ */
+export function isPureComponent(node: TSESTree.Node, context: RuleContext) {
+  const pragma = getPragmaFromContext(context);
+
+  const sourceCode = context.getSourceCode();
+
+  if ("superClass" in node && node.superClass) {
+    const text = sourceCode.getText(node.superClass);
+
+    // eslint-disable-next-line security/detect-non-literal-regexp
+    return new RegExp(`^(${pragma}\\.)?PureComponent$`, "u").test(text);
+  }
+
+  return false;
+}
 
 export function componentCollectorLegacy(context: RuleContext) {
   const components = new Map<string, ExRClassComponent>();
@@ -30,7 +76,7 @@ export function componentCollectorLegacy(context: RuleContext) {
     const id = getClassIdentifier(node);
     const key = uid.rnd();
     const flag = isPureComponent(node, context)
-      ? ExRClassComponentFlag.Pure
+      ? ExRClassComponentFlag.PureComponent
       : ExRClassComponentFlag.None;
     components.set(
       key,
@@ -42,6 +88,8 @@ export function componentCollectorLegacy(context: RuleContext) {
         // TODO: get displayName of class component
         displayName: O.none(),
         flag,
+        // TODO: get methods of class component
+        methods: [],
         node,
       },
     );
